+++ sequana/lazy.py
@@ -2,12 +2,14 @@
 
 from sequana.lazyimports import LazyImport
 
-# matplotlib
+# lazy imports 
 pylab = LazyImport('pylab')
 numpy = LazyImport('numpy')
 scipy = LazyImport('scipy')
 scipy_stats = LazyImport('scipy.stats')
 pandas = LazyImport('pandas')
+reports = LazyImport('reports')
+vcf = LazyImport("vcf")
 
 def enabled():
     "Returns ``True`` if LazyImports are globally enabled"