+++ sequana/reporting/__init__.py
@@ -1,1 +1,0 @@
-from .report_bam import BAMReport